refactor: ViewModel 레이어의 비즈니스 로직을 UseCase로 분리
